+++ frontend/api/_lib/types/archive.type.ts
@@ -1,0 +1,10 @@
+export type Archive = {
+  id: string
+  name: string            // "[MM]-[YYYY]"
+  month: number
+  year: number
+  count: number
+  debit: number
+  credit: number
+  transactionsDbId: string | null
+}